requirements.txt
# Flask Core
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Limiter==3.5.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Cache and Sessions
redis==5.0.0
Flask-Session==0.5.0

# Security
Werkzeug==2.3.7
WTForms==3.0.1
email-validator==2.0.0

# Monitoring and Metrics
prometheus-flask-exporter==0.23.0
psutil==5.9.5
structlog==23.1.0

# HTTP and API
requests==2.31.0
urllib3==2.0.5

# Utilities
python-dotenv==1.0.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3

# Date and Time
python-dateutil==2.8.2

# Development and Testing (optional)
pytest==7.4.2
pytest-flask==1.2.0
coverage==7.3.2
black==23.9.1
flake8==6.1.0

# Production WSGI Server
gunicorn==21.2.0

# Additional utilities
uuid==1.30
json5==0.9.14